correct "set" commands. use "source" instead of "sh".
add lxc-destroy to cleanup script.
JIRA: COPPER-2

+++ components/congress/joid/clean_congress.sh
@@ -18,17 +18,24 @@
 # LXC container in the OPNFV Controller node.
 # Presumably something has failed, and any record of the Congress feature
 # in OpenStack needs to be removed, so you can try the install again.
-# Most of the items removed by this script were created in install_congress_5.sh
 
 source ~/admin-openrc.sh <<EOF
 openstack
 EOF
+source ~/env.sh
+set -x echo on
+juju ssh ubuntu@node1-control "sudo lxc-stop --name juju-trusty-congress; sudo lxc-destroy --name juju-trusty-congress; exit"
 
 # Delete Congress user
 export CONGRESS_USER=$(openstack user list | awk "/ congress / { print \$2 }")
-openstack user delete $CONGRESS_USER
+if [ "$CONGRESS_USER" != "" ]; then
+  openstack user delete $CONGRESS_USER
+fi
 
 # Delete Congress service
 export CONGRESS_SERVICE=$(openstack service list | awk "/ congress / { print \$2 }")
-openstack service delete $CONGRESS_SERVICE
+if [ "$CONGRESS_SERVICE" != "" ]; then
+  openstack service delete $CONGRESS_SERVICE
+fi
+set -x echo off
